% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angry_face.R
\name{angry_face}
\alias{angry_face}
\title{An Angry Face Function}
\usage{
angry_face()
}
\description{
This function allows you to print out an angry face to the console
}
\examples{
angry_face()
}
\keyword{faces}
