% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_face.R
\name{happy_face}
\alias{happy_face}
\title{A Happy Face Function}
\usage{
happy_face(eyebrows = FALSE)
}
\arguments{
\item{eyebrows}{Do you want eyebrows on your face? Defaults to FALSE}
}
\description{
This function allows you to print out a happy face to the console
}
\examples{
happy_face()
}
\keyword{faces}
