% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_face.R
\name{neutral_face}
\alias{neutral_face}
\title{A Neutral Face Function}
\usage{
neutral_face(eyebrows = FALSE)
}
\arguments{
\item{eyebrows}{Do you want eyebrows on your face? Defaults to FALSE}
}
\description{
This function allows you to print out a neutral face to the console
}
\examples{
neutral_face()
}
\keyword{faces}
