% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad_face.R
\name{sad_face}
\alias{sad_face}
\title{A Sad Face Function}
\usage{
sad_face(tears = FALSE)
}
\arguments{
\item{tears}{Do you want tears on your face? Defaults to FALSE}
}
\description{
This function allows you to print out a sad face to the console
}
\examples{
sad_face()
}
\keyword{faces}
